Se mejora uso de memoria para la factorizacion con divisiones triviales

+++ B_QSieve/B_sieve.sh
@@ -5,20 +5,27 @@ else
 	num=$2;
 fi
 ./B_QSieve "$1" "$num" "$3" "$4"
-wait
+exit_status=$?;
+echo $exit_status
+wait;
+if [[ "$exit_status" == "1" ]];then
+	exit 1;
+fi
 /home/ingesis/cado-nfs-2.3.0/cado-nfs-2.3.0/build/debian9/linalg/bwc/mf_scan --ascii-in mfile=matrix.txt --binary-out ofile=matrix.bin --freq
 wait
 cp matrix.bin matrix.rw.bin matrix.cw.bin /tmp/.
 wait
 rm -r /tmp/c60/
 wait
-/home/ingesis/cado-nfs-2.3.0/cado-nfs-2.3.0/build/debian9/linalg/bwc/bwc.pl :complete thr=2 m=64 n=64 nullspace=left interval=100 matrix=/tmp/matrix.bin wdir=/tmp/c60/c60.bwc interleaving=0
+echo "Solucionando Matriz...";
+/home/ingesis/cado-nfs-2.3.0/cado-nfs-2.3.0/build/debian9/linalg/bwc/bwc.pl :complete thr=2 m=64 n=64 nullspace=left interval=100 matrix=/tmp/matrix.bin wdir=/tmp/c60/c60.bwc interleaving=0 > outputbwc.txt
 wait
 cp /tmp/c60/c60.bwc/K.sols0-64.0.txt /home/ingesis/prsa/B_QSieve/.
 wait
 ./readKsol.sh K.sols0-64.0.txt > vec.txt
 wait
 ./productPoli.sh "$num"
-
-
-
+wait
+cat salidap.txt
+wait
+rm polinomio.txt
